# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "1.0.0"
  description: Sakani Dz - Real Estate & Tourism Platform API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VendorSummary:
      type: object
      required: [id, firstName, lastName, phone]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        isVerified:
          type: boolean
    Vendor:
      type: object
      required: [id, firstName, lastName, phone, email, status, allowedPropertyTypes, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: string
        email:
          type: string
        code:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, renewal_required, blocked, rejected]
        allowedPropertyTypes:
          type: array
          items:
            type: string
        isVerified:
          type: boolean
        lastAccessAt:
          type: ["string", "null"]
          format: date-time
        blockedAt:
          type: ["string", "null"]
          format: date-time
        subscriptionExpiresAt:
          type: ["string", "null"]
          format: date-time
        totalViews:
          type: integer
        totalProperties:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Property:
      type: object
      required: [id, vendorId, type, price, wilaya, status, photos, equipment, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        vendor:
          $ref: "#/components/schemas/VendorSummary"
        type:
          type: string
          enum: [apartment, villa]
        apartmentType:
          type: ["string", "null"]
          enum: [F1, F2, F3, F4, F5, null]
        description:
          type: ["string", "null"]
        equipment:
          type: array
          items:
            type: string
        price:
          type: number
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        photos:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [available, rented]
        wilaya:
          type: string
        siteId:
          type: ["integer", "null"]
        views:
          type: integer
        isFeatured:
          type: boolean
        createdAt:
          type: string
          format: date-time
    Site:
      type: object
      required: [id, vendorId, name, wilaya, photos, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        vendor:
          $ref: "#/components/schemas/VendorSummary"
        name:
          type: string
        description:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        photos:
          type: array
          items:
            type: string
        wilaya:
          type: string
        propertyCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    VendorStats:
      type: object
      required: [totalProperties, totalViews, totalSites, availableProperties, rentedProperties, viewsThisWeek, viewsThisMonth]
      properties:
        totalProperties:
          type: integer
        totalViews:
          type: integer
        totalSites:
          type: integer
        availableProperties:
          type: integer
        rentedProperties:
          type: integer
        viewsThisWeek:
          type: integer
        viewsThisMonth:
          type: integer
...
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # â”€â”€ List responses â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
    PropertyListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Property"
        pagination:
          $ref: "#/components/schemas/Pagination"
    SiteListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Site"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    VendorRegistrationInput:
      type: object
      required: [firstName, lastName, phone, email, propertyType, wilaya]
      properties:
        firstName:
          type: string
          minLength: 2
        lastName:
          type: string
          minLength: 2
        address:
          type: string
        phone:
          type: string
        email:
          type: string

        propertyType:
          type: string
          enum: [apartment, villa, site]
        apartmentType:
          type: string
          enum: [F1, F2, F3, F4, F5]
        latitude:
          type: number
        longitude:
          type: number
        price:
          type: number
        photos:
          type: array
          items:
            type: string
        description:
          type: string
        wilaya:
          type: string
    VendorUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        address:
          type: string
        phone:
          type: string
...
paths:
...
  /vendors/register:
    post:
      operationId: registerVendor
      tags: [vendors]
      summary: Submit a vendor registration request (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorRegistrationInput"
      responses:
        "201":
          description: Registration submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/me:
    get:
      operationId: getVendorProfile
      tags: [vendors]
      summary: Get authenticated vendor profile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    patch:
      operationId: updateVendorProfile
      tags: [vendors]
      summary: Update vendor profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/me/properties:
    get:
      operationId: listVendorProperties
      tags: [vendors]
      summary: List vendor's own properties
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyListResponse"
  /vendors/me/sites:
    get:
      operationId: listVendorSites
      tags: [vendors]
      summary: List vendor's own sites
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteListResponse"
  /vendors/me/stats:
    get:
      operationId: getVendorStats
      tags: [vendors]
      summary: Get vendor dashboard statistics
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorStats"

  # â”€â”€ Admin â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
...